+++ src/logic/model/BookEvaluation.java
@@ -9,20 +9,20 @@ import logic.model.users.User;
  */
 public class BookEvaluation {
 	
-	private double rating;
+	private int rating;
 	private String title;
 	private String body;
 	
 	private User author;
 	private Book relatedBook;
 	
-	public BookEvaluation(User author, Book relatedBook, double rating) {
+	public BookEvaluation(User author, Book relatedBook, int rating) {
 		this.author = author;
 		this.relatedBook = relatedBook;
 		this.rating = rating;
 	}
 
-	public double getRating() {
+	public int getRating() {
 		return rating;
 	}
